This patch provides an initial C++ RCC implementation: make new Worker=foo.rcc Language=c++ "externals" attribute now works on application assemblies like hdl assemblies buffersize attribute on spec ports can refer to other ports by name ocpirun has -w arg to specify worker names for instances getPort for Applications can now supply connection parameters
components/Makefile: add two test components for HDL properties and C++ worker
core/application/include/OcpiApplication.h: moved some functionality down to OU::Assembly where it belongs
-- and allowed the getPort API to take connection parameters
core/application/include/OcpiApplicationApi.h: ditto
core/application/src/OcpiApplication.cxx: ditto
-- plus ignore accessibility for parameter properties - we know what they are
core/application/src/ocpirun_main.cxx: add instance filtering based on worker name
core/container/ctests/bin/include/test_utilities.h: moved some RCC global types into the proper namespace
-- and added a slot to the RCCDisptach structure
core/container/ctests/workers/include/ConsumerWorker.h: ditto
core/container/ctests/workers/include/LoopbackWorker.h ditto
core/container/ctests/workers/include/ProdWorker.h: ditto
core/container/ctests/workers/include/UtGenericLoopbackWorkers.h: ditto
core/container/ctests/workers/include/UtZeroCopyIOWorkers.h: ditto
core/container/ctests/workers/src/CosumerWorker.c: ditto
core/container/ctests/workers/src/LoopbackWorker.c: ditto
core/container/ctests/workers/src/ProdWorker.c: ditto
core/container/ctests/workers/src/UtGenericLoopBackWorkers.c: ditto
core/container/ctests/workers/src/UtZeroCopyIOWorkers.c: ditto
core/container/hdl/src/NetDriver.cxx: moved "info" messages into "debug"
core/dataplane/rdma_driver_interface/interfaces/src/DtSharedMemoryInternal.cxx: ditto
core/dataplane/rdma_driver_interface/interfaces/src/DtTransferInternal.cxx: ditto
core/dataplane/transport/impl/src/OcpiTransport.cxx: ditto
core/container/interfaces/include/OcpiContainerPort.h: change unsigned to size_t
core/container/interfaces/src/OcpiContainerPort.cxx: debug aids for buffer size madness
core/dataplane/transport/impl/src/OcpiPort.cxx: ditto
core/container/rcc/include/OcpiRccApplication.h: changed for C++ workers, preparing for contorl proxies
core/container/rcc/include/OcpiRccPort.h: ditto
core/container/rcc/include/OcpiRccWorker.h: ditto
core/container/rcc/include/RCC_Worker.h: ditto
core/container/rcc/src/OcpiRccApplication.cxx: ditto
core/container/rcc/src/OcpiRccWorker.cxx: ditto
core/library/include/OcpiLibraryAssembly.h: move impl filtering from app to lib level, add worker name filtering
core/library/include/OcpiLibraryManager.h: ditto
core/library/src/OcpiLibraryAssembly.cxx: ditto
-- Also support "externals" in app assemblies
core/library/src/OcpiLibraryManager.cxx: ditto
core/local/util/assembly/README.txt: update notes
core/local/util/assembly/include/OcpiUtilAssembly.h: moved instance params from app layer to here
core/local/util/assembly/src/OcpiUtilAssembly.cxx: ditto
core/local/util/property/include/OcpiPValue.h: added new merging PValueList constructor
core/local/util/property/src/OcpiUtilPValue.cxx: ditto
core/local/util/property/include/OcpiUtilImplementation.h: add non-throwing property lookup
core/local/util/property/include/OcpiUtilPort.h: added support for buffersizeport
-- where the buffer size of a port is derived from the size at another port for adapters
core/local/util/property/src/OcpiUtilImplementation.cxx: ditto
core/local/util/property/src/OcpiUtilPort.cxx: ditto
core/local/util/property/include/OcpiUtilProperty.h: remember which properties are impl-only
hdl/adapters/wsi_16_to_32.hdl/wsi_16_to_32.xml: use new buffersizeport feature
hdl/adapters/wsi_32_to_16.hdl/wsi_32_to_16.xml: ditto
jkamd_env_linux_x86_64.sh: updaet redhat env setup to be more like centos6
tools/cdk/examples/copy/copy.cxx: more testing and showing more features.
tools/cdk/include/hdl/hdl-targets.mk: added zzed for "pure zynq"
tools/cdk/include/lib.mk: allow specfiles in 'make new' without xml suffix being mentioned.
tools/cdk/include/rcc/default.mk: c++ update
tools/cdk/include/rcc/rcc-worker.mk: ditto
tools/cdk/include/util.mk: ditto
tools/local/tester/file_io_workers/include/fileSink.h: accomodate RCC namespace changes
tools/ocpigen/include/wip.h: support buffersizeport feature, and some rcc C++ issues
tools/ocpigen/src/assembly.cxx: ditto plus allow mixing "externals" with unnamed port connections
tools/ocpigen/src/hdl-assembly.cxx: better error messages
tools/ocpigen/src/ocpigen_main.cxx: emitSkelRCC is now a worker method.
tools/ocpigen/src/parameters.cxx: implement parameter matching/search in HDL assemblies
tools/ocpigen/src/parse.cxx: implement buffersizeport feature
tools/ocpigen/src/rcc.cxx: C++ code gen implementation

+++ core/container/interfaces/include/OcpiContainerPort.h
@@ -65,7 +65,7 @@ namespace OCPI {
      * Port data structure
      *********************************/  
     const unsigned DEFAULT_NBUFFERS = 2;
-    const unsigned DEFAULT_BUFFER_SIZE = 2*1024;
+    const size_t DEFAULT_BUFFER_SIZE = 2*1024;
     const unsigned BUFFER_ALIGNMENT = 16;
     class PortData
     {
